Added list swap and duplication
These work (hasve been tested).

Also removed leading + on integers and floats because it made it confusing when
writing something like:

1 1+1+

If + is part of the integer, you get [1,2] on the stack when you might expect 3.
I don't know why you'd want to assert +, so only a leading - is valid now and
writing the above gives [3] as probably expected.