Added macros
Pure substitution and substitution with group references works.

We have decided tentatively on "ṁ" being the macro leader but maybe this
will change. Of course macros can use any letter in the language. How
this works in the end is not yet fully defined: if macros are only
expanded once, then using symbols already in the language is allowed as
long as they expand to meaningful symbols; if macros are expanded to
exhaustion, you can't do that unless you have some way of "escaping" the
symbol.

For now macros are only expanded once. Perhaps in the future,
expansion-to-exhaustion can be enabled per macro: some macros use it,
others not.